quara/requirements.txt
# QuARA - Quantitative Academic Research Agent
# Core dependencies for multi-agent research framework

# Core language model and agents
openai>=1.12.0
anthropic>=0.17.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-anthropic>=0.1.0
langgraph>=0.0.40

# Multi-agent frameworks
crewai>=0.28.0
autogen-agentchat>=0.2.0

# Memory and vector storage
chromadb>=0.4.18
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Data science and analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Causal inference
dowhy>=0.9.1
econml>=0.14.0
causalml>=0.14.0

# Data cleaning and validation
great-expectations>=0.17.0
pandas-profiling>=3.6.0

# Web search and APIs
duckduckgo-search>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
arxiv>=1.4.8
biopython>=1.81  # For PubMed API

# Sandbox and execution
docker>=6.1.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Async and parallel processing
asyncio
aiohttp>=3.8.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Logging and observability
loguru>=0.7.0
wandb>=0.15.0
langsmith>=0.0.40

# File handling
pathlib
uuid
json5>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0